Textarea FIFO to own class
