+++ java/test/jmri/jmrix/can/cbus/swing/PackageTest.java
@@ -17,6 +17,7 @@ import org.slf4j.LoggerFactory;
     jmri.jmrix.can.cbus.swing.simulator.PackageTest.class,
     jmri.jmrix.can.cbus.swing.CbusFilterFrameTest.class,
     jmri.jmrix.can.cbus.swing.CbusFilterPanelTest.class,
+    jmri.jmrix.can.cbus.swing.TextAreaFIFOTest.class,
     BundleTest.class
 })
